Refactor: * Merge option and result into prelude * Split result tests into different modules
